feat: refactor project into a python package; add support for env var from zshrc or bashrc
