--- __init__.py
@@ -1,5 +1,0 @@
-"""
-Context Switcher Assistant - A tool to help manage context switching between different work sessions.
-"""
-
-__version__ = "0.1.0"
