+++ context_switcher/__init__.py
@@ -1,0 +1,5 @@
+"""
+Context Switcher Assistant - A tool for generating AI-enhanced git commit summaries
+"""
+
+__version__ = "0.1.0"
