+++ requirements.txt
@@ -1,3 +1,2 @@
 openai>=1.0.0
-python-dotenv>=0.19.0
 colorama>=0.4.6
